Update servers tests
